+++ ecc/blocks/img-upload-component/controller.js
@@ -1,7 +1,7 @@
 /* eslint-disable no-unused-vars */
 import { deleteImage, getEventImages, uploadImage } from '../../scripts/esp-controller.js';
 import { LIBS } from '../../scripts/scripts.js';
-import { getFilteredCachedResponse } from '../../scripts/event-data-handler.js';
+import { getFilteredCachedResponse } from '../event-creation-form/data-handler.js';
 
 const { createTag } = await import(`${LIBS}/utils/utils.js`);
